Begun integration of VoCore2 module
Took over some configuration files from the official VoCore source repository.

Currently there is an image created but it's not able to boot up properly.
There is something missing arround spi flash. The created kernel is not able to mount the rootfs.

I reckon that there is something missing for detecting the spi flash on the vocore module. Or some configuration or kernel patches are missing.

The kernel patches in repository of the vocore module have to be transfered to this branch.